requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
slowapi==0.1.9
python-multipart==0.0.9
pydantic-settings==2.4.0

# tests / qa
pytest==8.3.3
pytest-cov==5.0.0
ruff==0.6.9
bandit==1.7.9
pip-audit==2.7.3
detect-secrets==1.5.0
httpx==0.27.2

# IA
openai>=1.43.0